Extract GCD tests into dedicated test file
Moves tests for `compute_gcd` and `prime_factorization_for_gcd` from the general `Tests/test_DiscreteMath.py` file into a dedicated `Tests/test_gcd.py` file to improve test organization and modularity. Added proper sys.path handling to the new test file to resolve internal math module imports.

All tests pass successfully.
