+++ Tests/test_gcd.py
@@ -1,0 +1,66 @@
+import os
+import sys
+
+import pytest
+
+# Add root directory to path to allow "Math.Discrete_Math..." imports
+root_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
+if root_dir not in sys.path:
+    sys.path.insert(0, root_dir)
+
+# the code has imports assuming "Math" is the root in some cases, so let's add it too
+math_dir = os.path.abspath(os.path.join(root_dir, 'Math'))
+if math_dir not in sys.path:
+    sys.path.insert(0, math_dir)
+
+from Math.Discrete_Math.Number_Theory.gcd import compute_gcd, prime_factorization_for_gcd
+
+
+def test_compute_gcd_coprime():
+    assert compute_gcd(17, 19) == 1
+    assert compute_gcd(8, 9) == 1
+
+
+def test_compute_gcd_multiples():
+    assert compute_gcd(10, 5) == 5
+    assert compute_gcd(5, 10) == 5
+    assert compute_gcd(12, 36) == 12
+
+
+def test_compute_gcd_identical():
+    assert compute_gcd(7, 7) == 7
+    assert compute_gcd(100, 100) == 100
+
+
+def test_compute_gcd_common_factors():
+    assert compute_gcd(48, 18) == 6
+    assert compute_gcd(54, 24) == 6
+
+
+def test_compute_gcd_errors():
+    with pytest.raises(ValueError, match="Both numbers must be positive."):
+        compute_gcd(0, 5)
+    with pytest.raises(ValueError, match="Both numbers must be positive."):
+        compute_gcd(5, 0)
+    with pytest.raises(ValueError, match="Both numbers must be positive."):
+        compute_gcd(-5, 5)
+
+
+@pytest.mark.parametrize("n, expected", [
+    # Test composite numbers
+    (12, [2, 2, 3]),
+    (60, [2, 2, 3, 5]),
+    (100, [2, 2, 5, 5]),
+
+    # Test prime numbers
+    (2, [2]),
+    (7, [7]),
+    (13, [13]),
+
+    # Test edge cases
+    (1, []),
+    (0, []),
+    (-5, [])
+])
+def test_prime_factorization_for_gcd(n, expected):
+    assert prime_factorization_for_gcd(n) == expected
